+++ src/components/cookie/CookieNotice.jsx
@@ -8,7 +8,7 @@ import { injectIntl } from 'react-intl';
 
 import CookieSettings from './CookieSettings';
 
-import createINTL from '../../util/createINTL';
+import createInternationalization from '../../util/createInternationalization';
 
 const Link = styled.span`
   text-align: center;
@@ -57,7 +57,7 @@ class CookieNotice extends React.Component {
   }
 
   render() {
-    const intl = createINTL(this.props.intl);
+    const intl = createInternationalization(this.props.intl);
     const { locale } = this.props;
 
     return (