+++ src/components/form/RegisterBusinessForm.jsx
@@ -23,7 +23,7 @@ import {
   REGISTER_BUSINESS_FORM_NAME,
 } from '../../constants';
 
-import createINTL from '../../util/createINTL';
+import createInternationalization from '../../util/createInternationalization';
 import encodeFormState from '../../util/encodeFormState';
 
 const FormContainer = styled.div`
@@ -94,7 +94,7 @@ class RegisterBusinessForm extends React.Component {
   }
 
   validatePageData() {
-    const intl = createINTL(this.props.intl);
+    const intl = createInternationalization(this.props.intl);
     const errors = {};
     let isValid = true;
 
@@ -304,7 +304,7 @@ class RegisterBusinessForm extends React.Component {
   }
 
   render() {
-    const intl = createINTL(this.props.intl);
+    const intl = createInternationalization(this.props.intl);
     const { locale } = this.props;
 
     return (