+++ src/components/form/RegisterPersonForm.jsx
@@ -22,7 +22,7 @@ import {
   REGISTER_PERSON_FORM_NAME,
 } from '../../constants';
 
-import createINTL from '../../util/createINTL';
+import createInternationalization from '../../util/createInternationalization';
 import encodeFormState from '../../util/encodeFormState';
 
 const FormContainer = styled.div`
@@ -88,7 +88,7 @@ class RegisterPersonForm extends React.Component {
   }
 
   validatePageData() {
-    const intl = createINTL(this.props.intl);
+    const intl = createInternationalization(this.props.intl);
     const errors = {};
     let isValid = true;
 
@@ -242,7 +242,7 @@ class RegisterPersonForm extends React.Component {
   }
 
   render() {
-    const intl = createINTL(this.props.intl);
+    const intl = createInternationalization(this.props.intl);
     const { locale } = this.props;
 
     return (