+++ src/templates/Authors.jsx
@@ -15,7 +15,7 @@ import Rule from '../components/Rule';
 import SchemedImage from '../components/SchemedImage';
 import Theme from '../components/Theme';
 
-import createINTL from '../util/createINTL';
+import createInternationalization from '../util/createInternationalization';
 
 const Separator = styled.div`
   display: flex;
@@ -145,7 +145,7 @@ const propTypes = {
 const defaultProps = { children: undefined };
 
 function Page({ data, pageContext }) {
-  const intl = createINTL(useIntl());
+  const intl = createInternationalization(useIntl());
 
   const { edges: authors } = data.allContentfulAuthor;
   const { locale, pageID } = pageContext;