+++ src/templates/Categories.jsx
@@ -13,7 +13,7 @@ import LocalizedLink from '../components/link/LocalizedLink';
 import Rule from '../components/Rule';
 import Theme from '../components/Theme';
 
-import createINTL from '../util/createINTL';
+import createInternationalization from '../util/createInternationalization';
 
 const Separator = styled.div`
   display: flex;
@@ -90,7 +90,7 @@ const propTypes = {
 const defaultProps = { children: undefined };
 
 function Page({ data, pageContext }) {
-  const intl = createINTL(useIntl());
+  const intl = createInternationalization(useIntl());
 
   const { edges: categories } = data.allContentfulCategory;
   const { edges: posts } = data.allContentfulBlogPost;