+++ src/templates/Category.jsx
@@ -16,7 +16,7 @@ import LocalizedLinkButton from '../components/link/LocalizedLinkButton';
 import Rule from '../components/Rule';
 import Theme from '../components/Theme';
 
-import createINTL from '../util/createINTL';
+import createInternationalization from '../util/createInternationalization';
 
 const Post = styled.article`
   margin: 2em ${(props) => props.theme.layout.marginMobile};
@@ -111,7 +111,7 @@ const propTypes = {
 const defaultProps = { children: undefined };
 
 function Page({ data, pageContext }) {
-  const intl = createINTL(useIntl());
+  const intl = createInternationalization(useIntl());
 
   const { contentfulCategory: category } = data;
   const { edges: posts } = data.allContentfulBlogPost;