+++ src/util/createInternationalization.js
@@ -1,6 +1,6 @@
 // Copyright (c) 2021 Visiosto oy
 // Licensed under the MIT License
 
-export default function createINTL(intlUseEffect) {
+export default function createInternationalization(intlUseEffect) {
   return (k, obj) => intlUseEffect.formatMessage({ id: k }, obj);
 }